Add quarterly idle-balance email cron
