+++ vercel.json
@@ -19,6 +19,10 @@
     {
       "path": "/api/weekly-digest",
       "schedule": "0 17 * * 1"
+    },
+    {
+      "path": "/api/email-idle-balances",
+      "schedule": "0 0 1 2,5,8,11 *"
     }
   ]
 }
